+++ src/components/CardComponentTwo.tsx
@@ -1,0 +1,28 @@
+import { CardGroup, Card } from "react-bootstrap";
+
+type CardComponentTwoProps = {
+    imgSrc ?: string;
+    cardTitle ?: string; 
+    cardText ?: string;
+    cardFooter ?: string;
+
+}
+
+export const CardComponentTwo = (props : CardComponentTwoProps) => {
+    return (
+        // <CardGroup>
+        <Card className="p-2 m-2 custom-card-2">
+          <Card.Img variant="top" src={props.imgSrc} />
+          <Card.Body>
+            <Card.Title>{props.cardTitle}</Card.Title>
+            <Card.Text>
+                {props.cardText}              
+            </Card.Text>
+          </Card.Body>
+          <Card.Footer>
+            <small className="text-muted">{props.cardFooter}</small>
+          </Card.Footer>
+        </Card>
+    //   </CardGroup>
+    );
+}