+++ src/components/HeadingComponent.tsx
@@ -1,7 +1,7 @@
 export const HeadingComponent = () => {
     return (
-        <div className="text-center" style={{ margin: '80px'}}>
-            <h1 className="m-4">Heading One</h1>
+        <div className="text-center" style={{ margin: '40px'}}>
+            <h1 className="m-md-4">Heading One</h1>
             <hr style={{ margin: 'auto', width: '20px' }} />
         </div>